Added animations and finished the team/driver scraping
so basically next  up im gonna utilise the  web scraping ive just completed. to be completely honest this step was so extra but i know that i would have been sad if i just copy and pasted and didnt do it the way i originally wanted too. I love the solution to the changing picture though i though it was so neat to use a usestate which im coming to realise is super important in front end development. I also a, realising front end development is pretty cool,  however it still feels awesome to be able to do the backend and really understand what is going on.